backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
opencv-python
python-docx
PyPDF2
sentence-transformers
chromadb
openai
reportlab
whisper
celery
redis
python-dotenv
requests
sqlalchemy
asyncpg
opencv-python
ftfy
torch>=1.10
transformers
tqdm
Pillow
git+https://github.com/openai/CLIP.git